RecoverableJournalQueuingStrategy: new class
RecoverableJournalQueuingStrategy provides the same behaviour as the
JournalQueuingStrategy but providing recovery features. It is thus
more robust and resistent to failures.

The RecoverableJournalQueuingStrategy can recover from the following
sitations:

- partially written buffer (power supply failure)
- buffer not append to index (lost, power supply failure)
- referenced buffer becomes broken (I/O issues)

Recovery uses file recover.tmp for recovering buffers to avoid collisions
with data.tmp used by the regular use of JournalQueuingStrategy.

The RecoverableJournalQueuingStrategy instance for GWS is initially
disabled. Enable it by setting:

 exporters.gws.tmpStorage.impl = recoverable

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>