BelkinWemoDeviceManager: migrate to generic handleListen()
The discovery process of the BelkinWemoDeviceManager is already asynchonous
via the BelkinWemoSeeker class. Thus, we just adapt the BelkinWemoSeeker to
AsyncWork interface.

We can share the BelkinWemoSeeker instance among calls to startDiscovery()
as it is guaranteed that the startDiscovery() is called and progressing
only once at a time and no other startDiscovery() is initiated until
the previous discovery finishes. This allows to delete m_seekerMutex.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>