PhilipsHueDeviceManager: migrate to generic handleListen()
The discovery process of the PhilipsHueDeviceManager is already asynchonous
via the PhilipsHueSeeker class. Thus, we just adapt the PhilipsHueSeeker to
AsyncWork interface.

We can share the PhilipsHueSeeker instance among calls to startDiscovery()
as it is guaranteed that the startDiscovery() is called and progressing
only once at a time and no other startDiscovery() is initiated until
the previous discovery finishes. This allows to delete m_seekerMutex.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>