DeviceManager: copy resulted set of device IDs
This patch seems to fix segfaulting of certain situations while
handling the unpair command. It is not very clear why this happens
as it should not happen. My understanding:

1. unpair is some instance represented as AsyncWork<set<DeviceID>>::Ptr
2. unpair->result() returns Nullable<set<DeviceID>>
3. we grab a reference to this temporary copy of the set<DeviceID>
4. we return a copy of the temporary set<DeviceID>
...
5. accessing the returned set<DeviceID> (which should to be a valid
   copy) might cause segfault.

The segfault cannot be seen on the actual code base. It is necessary
to apply unpair updates to the PressureSensorManager and unpair the
pressure sensor.

Fixes: 685787002197 ("DeviceManager: implement handleUnpair()")
Reported-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>