BLESmartDeviceManager: new class
The class implements the work with BLE smart devices. This means
seeking for BLE smart devices on the network, modifying state of
the devices and obtaining data from the devices.

There are two ways to obtain data. The first way is when the
manager requests the data from the device at certain time interval.
The second way is when the device sends data and the manager immediately
processes it. This allows to send data immediately after the data are
recieved to the server, which is convenient for the motion sensor and
door/window sensor.

The BLESmartDeviceManager does not use the DeviceStatusFetcher but
it can be modified that way easily.

Duration of the discovery process depends on the parameter scanTimeout.
Its value defines the time period of 1 discovery iteration. Thus, it
takes N * scanTimeout to perform discovery.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>