JSONSensorDataFormatter: output doubles in fixed format always
Sometimes, it happens, that ostream prints double in the scientific
notation like 1.67772e+07. This can however lead to loosing of
precision which is not wanted in certain situations.

Example:

 out << ((double) 16777215)

prints

 1.67772e+07

which is however 16777200.

Note that in RecoverableJournalQueuingStrategyTest, we had to change
the cut point of (former raw_1e90a60) raw_b2d3703.substr(0, 168). The
size 168 would not keep all the first record because of the added zeros.
A new magic number is 179 which is more than enough (intentionally not
exact).

Fixes: abf6c73cfa29 ("JSONSensorDataFormatter: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>