OZWNetwork: implement reset of controllers
The OZWNetwork is able to reset controllers according to the configuration.
Each controller can be reset only once since the last start of the OZWNetwork.
Controllers to be reset are specified via their home IDs. A list of home IDs
can be passed into the OZWNetwork::setControllersToReset() property. When such
controller appears in the Z-Wave network, the reset command is issued for it.

Use the configuration option like:

 zwave.controllers.reset = E6431D85

to configure the OZWNetwork to reset such controller when it is plugged in.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>