types-mapping.xml: move type mappings from GenericZWaveMapperRegistry
The mapping between Z-Wave command classes and BeeeOn types can be defined
in a configuration file to avoid hard-wiring it. Thus, move the mappings
into a file types-mapping.xml and point the GenericZWaveMapperRegistry to
parse it.

The GenericZWaveMapperRegistryTest has to be fixed to load mapping for
tests because it is not built-in anymore.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>