GenericZWaveMapperRegistry: implement loading of type mappings
The mapping of Z-Wave types to BeeeOn types can now be provided from
an external XML file. The file would override the hard-wired mappings
entirely.

A DTD file to validate a type-mappings.xml file was created.
A single types-mapping.xml file can potentially hold mappings for other
technologies as well if needed and defined in the future.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>