config: several Z-Wave types are converted to unknown
Several Z-Wave command classes are a kind of binary and because there
are no counterparts on the BeeeOn side yet, they have been converted
to the security_alert. This is however backwards incompatible.

Thus, all such Z-Wave command classes are now converted to unknown
which can be changed later and currently, it can be tuned on the
server side if needed.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>