HciInfoReporter: extract class from BluetoothAvailabilityManager
Reporting statistics about the HCI should be done independently from the
running device managers. Currently, only BluetoothAvailabilityManager could
report stats. The BLESmartDeviceManager does not implement that functionality.

Now on, the HciInfoReporter takes care of reporting HCI statistics and the
device managers are not involved in this process anymore. Statistics are
reported for all available Bluetooth devices.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>