ZWaveDeviceManager: implement set-value command
The method startSetValue() calls ZWaveNetwork::postValue() to set a value
on a given device. It is not very clear how to verify that value has been
truly set. According to some discussion in the OpenZWave community, it
seems that some devices might respond inaccurately or with a delay and
thus the reported Z-Wave value-changed events would not help us a lot.

Signed-off-by: Martin Hošala <xhosal00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>