ba-oscillation-detector: new python script
Python script for oscillation detection of bluetooth device presence.
The script connects to the MQTT broker from which receive the message
sent by the beeeon gateway. The oscillation is detected so that the
change of the presence of device occurs after less than 5 measurements.

Requires the Eclipse Paho MQTT Python client library:

 https://pypi.org/project/paho-mqtt/

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
[deduplicated description and executable permissions]
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>