DBusHciInterface::Device: merge WatchedDevice with Device
In order not to keep for watched devices both instances of
Device and WatchedDevice class, the Device class is expanded with
WatchedDevice functionality.

Due to this merge of WatchedDevice with Device are also changed
watch() and unwatch() methods that are now working with m_devices.
As each device can now be marked as watched or not-watched. The
WatchedDevice has no sense anymore and it is removed.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>