JablotronDeviceManager: remake to use JablotronController
The original JablotronDeviceManager was considerably simplified.
The communication logic is now hidden inside the JablotronController.
The JablotronDeviceManager implements the new DeviceManager API.

There is a significant modification that makes the JablotronDeviceManager
incompatible with the previous implementation. The AC-88 is not treated
as an active element anymore. Both halves of RC-86K are now managed as
a single device. Moreover, 3 new modules are introduced:

* PGX - on/off controllable
* PGY - on/off controllable (unclear how to pair it by software)
* Siren - on/off controllable and beeping controllable

The TX messages are send multiple-times based on the configured BackOff
strategy. The used strategy is RandomBackOff configured to generate
2 delays in range 200-500 ms, i.e. the TX message is sent 3 times.

The incompatibilities are reflected in the device names. The AC-88
is named AC-88 (sensor) and the RC-86K is named RC-86K (dual).

The JablotronDeviceManager now supports pairing by writing or erasing
the slots in the Jablotron dongle. Slots are not cached inside the
JablotronDeviceManager instances and are always read when needed.

The PGX, PGY and Siren controls are paired by sending the ENROLL:1 packet
when the target devices is in the appropriate learning mode. It was tested
on the AC-88 and JA-80L that were paired without any issues. The status
of controls cannot be obtained in any way. Thus, when started, the initial
values are assumed and used.

The DongleDeviceManager abstract class is not used anymore. Instead,
the HotplugListener is implemented directly as it is more suitable for
this case.

Unpairing may or may not (based on the configuration of unpairErasesSlot)
unregister the target gadget from its slot.

The JablotronDeviceManager's pairing cache is now synced by the configured
DeviceStatusFetcher. When the dongle is connected, the manager tries to
sync the registered slots with the pairing cache. If the DeviceStatusFetcher
delivers paired devices, the pairing cache is updated and if there is dongle
connected, it is synced the same way.

The JablotronDevice-like classes are not used anymore.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>