DeviceStatusFetcher: fix PrefixStatus::needsRequest()
The implementation of needsRequest() was insufficient. It was not
possible to determine whether we need to make a request just from
the m_successful state. Consider:

 If m_successful == false, is it possible to say that a request
 has been made?

No. Thus, we have introduced a variable m_started that stores
this information. If a request has started, we can still perform
a repeat but that is another story.

Without this fix, we fail with 2 or more handlers on:

 Assertion violation: !m_successful in file
 	"src/core/DeviceStatusFetcher.cpp", line 52

Fixes: ccce884771af ("DeviceStatusFetcher: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>