DeviceManager: introduce method waitRemoteStatus()
Becuase method deviceList() (and related) are marked deprecated and
some modules still use it, we should introduce a transition mechanism
that is less wierd and compatible with usage of the handleRemoteStatus()
method.

Now, those who call deviceList() can be switched to using of the
handleRemoteStatus() method easily by declaring

 BEEEON_OBJECT_CASTABLE(DeviceStatusHandler)

by registering with a DeviceStatusFetcher instance and by calling
waitRemoteStatus() instead of deviceList().

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>