Distinguish between Z-Wave support and OpenZWave support
The Z-Wave core support is independent from the OpenZWave library.
Thus, the OZWNotificationEvent is not available when OpenZWave is
not to be linked.

The macro HAVE_ZWAVE is now set when having the core Z-Wave support
and a new macro HAVE_OPENZWAVE is introduced when having the OpenZWave
library available.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>