config: check that only 1 gateway instance is running
Starting to gateway instances on a single machine is usualy a mistake
in production. The SingleInstanceChecker checks whether there is another
instance running on startup. If there is another one, it simply fails.

To start multiple instances, change the application.instance.mode to
"ignore". In case when an instance failed so terribly that it blocks
another instance to run, then choose the mode "recover" instead.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>