JablotronDeviceManager: implement startSearch by serial-number
The Turris Dongle does not provide any device discovery capabilities.
Thus, it is necessary to register the gadgets into slots manually.
We would like to be able to install new gadgets easily. The command
DeviceSearchCommand can be used for this purpose.

Searching a gadget by its serial number is implemented as a slot
registration request. If the given serial number is a valid gadget
address, such gadget is registered and reported as a new device.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>