FilesystemDeviceCacheTest: remove named mutexes on tear down
The FilesystemDeviceCache creates named mutexes sometimes.
On UNIX, the mutexes are linked to files. The Poco::NamedMutex
for UNIX creates the files in /tmp/<name>.mutex.

If the FilesystemDeviceCacheTest is executed under 2 different
system users account, one of the tests might fail because they
shared the same mutex. The mutex back-file might be inaccessible
under the other account. Thus, it is better to remove the mutexes
when a test finishes.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>