GWSConnectorImpl: new class
GWSConnectorImpl implements the actual communication with the remote
Gateway Server. It takes care of the low-level communication details
like keep-alive ping-pong, reconnections in case of a failure, queuing
of output messages.

The GWSConnectorImpl is built as a single thread trying to connect
to the Gateway Server. If it succeeds, the SocketReactor is created
to handle incoming traffic (more threads). On disconnect, the SocketReactor
is stopped and destroyed.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>