GWSOptimisticExporter: new class
Optimistic sensor data exporter that exports data without waiting
for confirmations. However, confirmations are expected to come.
If the configured limit of maximal exports to be done is reached,
exporting stops until any of them is confirmed. It can be quite fast
when exporing data but it does provide any other capabilities
(like persistance).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>