build: fix incomplete condition
We have to check both BLUETOOTH and ENABLE_BLE_SMART to enable
BLE tests. Otherwise, we can fail while trying to link against
a non-existing libBeeeOnBLE library.

The libBeeeOnBLE is built only when BLUETOOTH is true (the bluez
library is available). If the bluez library is missing, the
BLE tests were still about to built.

Fixes: 52f84f70f936 ("build: link Bluetooth LE separately as library")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>