poco-all: new package (version 1.9.0-all)
OpenWrt currently contains Poco in version 1.8.x. This would be
sufficient, however it does not include some important libraries
we need for beeeon-gateway.

Thus, we replace the OpenWRT Poco package by this one that is
named slightly different and provides all we need. The Poco
is of version 1.9.0 here.

Poco has a bug in the configure script that prevents it work
properly. Thus, the patch 100-configure.patch is required to
build.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>