BelkinWemoStandaloneDevice: make m_httpTimeout const
As m_httpTimeout is protected and can be changed from any inherited
class out of control. It is better and safer to mark it const. This
requires to change construction of BelkinWemoStandaloneDevice-like
classes a bit.

This change also improves the construction principle a bit because
now, the httpTimeout is given via constructor - from outside.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>