DeviceManager: handle operation mode of DeviceSetValueCommand
DeviceManager chooses the expected mode of setting a module value
based on the DeviceSetValueCommand::mode() property. There are thus
3 possible implementations of the startSetValue().

The default mode TRY_ONCE leads to call startSetValue() as usual.
This preserves compatibility with the previous behaviour.

The mode TRY_HARDER leads to call startSetValueTryHarder(). The
default implementation calls startSetValue(). In this way, it
does not change the behaviour of the operation.

The mode REPEAT_ON_FAIL leads to call startSetValueRepeatOnFail().
This is useful for blocking startSetValue() implementations. If an
IOException is thrown from startSetValue(), it is called again in
loop. However, in practice, this would have to be implemented better
and in a more specific way.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>