ConradDeviceManager: communicate via FHEMClient
The ConradDeviceManager do not use anymore conrad-interface for controling
Conrad devices. From now it uses FHEMClient to interact with FHEM server and
thus control the Conrad devices. This make the communication easier and lighter
because now the BeeeOn Gateway communicate directly with th FHEM server and not
over middle conrad-interface.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>